create src directories -- move src folders
