--- app/stories/pages/Home.stories.tsx
@@ -1,12 +1,0 @@
-import { ComponentStory, ComponentMeta } from '@storybook/react'
-
-import HomePage from '../../pages/index'
-
-export default {
-  title: 'Pages',
-  component: HomePage,
-} as ComponentMeta<typeof HomePage>
-
-const Template: ComponentStory<typeof HomePage> = () => <HomePage />
-
-export const Home = Template.bind({})
